:recycle: api: Move getV3CertificationCourseDetailsForAdministration injection to course
