+++ api/src/certification/course/domain/usecases/index.js
@@ -7,6 +7,7 @@ import * as competenceMarkRepository from '../../../../../lib/infrastructure/rep
 import * as answerRepository from '../../../../shared/infrastructure/repositories/answer-repository.js';
 import * as assessmentResultRepository from '../../../../shared/infrastructure/repositories/assessment-result-repository.js';
 import * as challengeRepository from '../../../../shared/infrastructure/repositories/challenge-repository.js';
+import * as competenceRepository from '../../../../shared/infrastructure/repositories/competence-repository.js';
 import { injectDependencies } from '../../../../shared/infrastructure/utils/dependency-injection.js';
 import { importNamedExportsFromDirectory } from '../../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
 import * as flashAlgorithmService from '../../../flash-certification/domain/services/algorithm-methods/flash.js';
@@ -19,9 +20,11 @@ import * as certificationReportRepository from '../../../shared/infrastructure/r
 import * as certificateRepository from '../../infrastructure/repositories/certificate-repository.js';
 import * as courseAssessmentResultRepository from '../../infrastructure/repositories/course-assessment-result-repository.js';
 import * as scoCertificationCandidateRepository from '../../infrastructure/repositories/sco-certification-candidate-repository.js';
+import * as v3CertificationCourseDetailsForAdministrationRepository from '../../infrastructure/repositories/v3-certification-course-details-for-administration-repository.js';
 /**
  * Using {@link https://jsdoc.app/tags-type "Closure Compiler's syntax"} to document injected dependencies
  *
+ * @typedef {competenceRepository} CompetenceRepository
  * @typedef {answerRepository} AnswerRepository
  * @typedef {challengeRepository} ChallengeRepository
  * @typedef {assessmentResultRepository} AssessmentResultRepository
@@ -37,8 +40,10 @@ import * as scoCertificationCandidateRepository from '../../infrastructure/repos
  * @typedef {pickChallengeService} PickChallengeService
  * @typedef {flashAlgorithmService} FlashAlgorithmService
  * @typedef {flashAlgorithmConfigurationRepository} FlashAlgorithmConfigurationRepository
+ * @typedef {v3CertificationCourseDetailsForAdministrationRepository} V3CertificationCourseDetailsForAdministrationRepository
  **/
 const dependencies = {
+  competenceRepository,
   answerRepository,
   challengeRepository,
   assessmentResultRepository,
@@ -54,6 +59,7 @@ const dependencies = {
   pickChallengeService,
   flashAlgorithmService,
   flashAlgorithmConfigurationRepository,
+  v3CertificationCourseDetailsForAdministrationRepository,
 };
 
 const path = dirname(fileURLToPath(import.meta.url));
@@ -66,7 +72,7 @@ const path = dirname(fileURLToPath(import.meta.url));
 const usecasesWithoutInjectedDependencies = {
   ...(await importNamedExportsFromDirectory({
     path: join(path, './'),
-    ignoredFileNames: ['index.js', 'get-v3-certification-course-details-for-administration.js'],
+    ignoredFileNames: ['index.js'],
   })),
 };
